requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
aniso8601==9.0.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.6
dataclasses-json==0.6.5
Flask==3.0.3
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
flask-swagger-ui==4.11.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.1.17
langchain-community==0.0.37
langchain-core==0.1.52
langchain-text-splitters==0.0.1
langsmith==0.1.54
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
pinecone-client==4.0.0
protobuf==4.25.3
pyarrow==16.0.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0
Pygments==2.18.0
pypdfium2==4.29.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.30
streamlit==1.34.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
Werkzeug==3.0.3
yarl==1.9.4
